feat(Aluno): adicionar rota de registro de Alunos & Exceptions

+++ Services/AlunoService.cs
@@ -1,6 +1,6 @@
-﻿using FitTrackAPI.Model;
+﻿using FitTrackAPI.Exceptions;
+using FitTrackAPI.Model;
 using FitTrackAPI.Repositories;
-using System.ComponentModel.DataAnnotations;
 
 namespace FitTrackAPI.Services
 {
@@ -20,47 +20,67 @@ namespace FitTrackAPI.Services
 
         public async Task<Aluno?> BuscarAlunoById(Guid id)
         {
+            var foundAluno = await _repository.GetByIdAsync(id);
+
+            if(foundAluno == null)
+            {
+                throw new NotFoundException("Aluno não encontrado!");
+            }
+
             return await _repository.GetByIdAsync(id);
         }
 
-        public async Task<IEnumerable<Aluno>> CadastrarAluno(string nome, string email)
+        public async Task<Aluno> CadastrarAluno(string nome, string email)
         {
             if(string.IsNullOrEmpty(nome))
             {
-                throw new Exception("Nome inválido!");
+                throw new ValidationException("Nome inválido!");
             }
 
             if(string.IsNullOrEmpty(email))
             {
-                throw new Exception("Email inválido!");
+                throw new ValidationException("Email inválido!");
             }
             
             var novoAluno = new Aluno(nome, email);
             await _repository.AddAsync(novoAluno);
 
-            return await _repository.GetAllAsync();
+            return novoAluno;
         }
 
-        public async Task<IEnumerable<Aluno>> AtualizarAluno(Guid id, string nome, string email, bool ativo)
+        public async Task<Aluno> AtualizarAluno(Guid id, string nome, string email, bool ativo)
         {
             if (string.IsNullOrEmpty(nome))
             {
-                throw new Exception("Nome inválido!");
+                throw new ValidationException("Nome inválido!");
             }
 
             if (string.IsNullOrEmpty(email))
             {
-                throw new Exception("Email inválido!");
+                throw new ValidationException("Email inválido!");
+            }
+
+            var alunoSelected = await _repository.GetByIdAsync(id);
+
+            if(alunoSelected == null)
+            {
+                throw new NotFoundException("Usuário com esse id não existe");
             }
 
             await _repository.UpdateAsync(id, nome, email, ativo);
-            return await _repository.GetAllAsync();
+            return alunoSelected;
         }
 
-        public async Task<IEnumerable<Aluno>> DeletarAluno(Guid id)
+        public async Task DeletarAluno(Guid id)
         {
+            var alunoSelected = _repository.GetByIdAsync(id);
+
+            if(alunoSelected == null)
+            {
+                throw new NotFoundException("Usuário com esse id não existe");
+            }
+
             await _repository.DeleteAsync(id);
-            return await _repository.GetAllAsync();
         }
     }
 }
